task: move socket
