+++ backend/app.js
@@ -87,41 +87,12 @@ app.use(function(err, req, res) {
  */
 const port = process.env.PORT || '3000';
 
-const io = require('socket.io').listen(app.listen(port));
+const http = require('http').Server(app);
+const socket = require('./socket/socket');
+socket(http);
 
-let userId = 0;
-let connections = [];
-
-io.sockets.on('connection', function(socket) {
-    connections.push(socket);
-    userId += 1;
-    socket.emit('start', { userId });
-    socket.on('message', async (data) => {
-        const user = await token.verify(data.token);
-        const chatroom = data.chatroom;
-
-        //store message to chat
-        chatroomService
-            .storeMessageToChatroom(data.message, data.user, chatroom)
-            .then((result) => {
-                //console.log('successfully stored ' + result);
-            })
-            .catch((err) => console.log(err));
-
-        connections.forEach((connectedSocket) => {
-            if (connectedSocket !== socket) {
-                connectedSocket.emit('message', {
-                    message: data.message,
-                    user: data.user,
-                });
-            }
-        });
-    });
-
-    socket.on('disconnect', () => {
-        const index = connections.indexOf(socket);
-        connections.splice(index, 1);
-    });
+http.listen(port, () => {
+    console.log('server is listening on port: ', port);
 });
 
 // export for testing
